// src/api/models/assignment-update-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AssignmentUpdateTO
 */
export interface AssignmentUpdateTO {
    /**
     * 
     * @type {string}
     * @memberof AssignmentUpdateTO
     */
    repositoryId: any;
    /**
     * 
     * @type {string}
     * @memberof AssignmentUpdateTO
     */
    username: any;
    /**
     * 
     * @type {string}
     * @memberof AssignmentUpdateTO
     */
    userId: any;
    /**
     * 
     * @type {string}
     * @memberof AssignmentUpdateTO
     */
    roleEnum: AssignmentUpdateTORoleEnumEnum;
}

/**
    * @export
    * @enum {string}
    */
export enum AssignmentUpdateTORoleEnumEnum {
    OWNER = 'OWNER',
    ADMIN = 'ADMIN',
    MEMBER = 'MEMBER',
    VIEWER = 'VIEWER'
}